+++ lib/example/models/money.dart
@@ -1,0 +1,5 @@
+class Money {
+  final int cents;
+
+  const Money(this.cents);
+}